requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
fastmcp>=0.9.0
google-genai>=0.1.0
faker>=20.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
httpx>=0.25.0
langchain-openai>=0.1.0
mcp-use>=0.1.0
